flask_web/requirements.txt
Flask==1.1.2
Pillow==8.1.2
fastai==2.2.7
